+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 History
 -------
 
+ - Add parameter maxsplit to strings.str2tuple
+ - Deprecate modules crypto and config
+
 **2020-07-12 (0.13.0)**
  - Add method Config.as_dict() in module config
  - Add exception DuplicateError in module config